% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting.R
\name{boosting}
\alias{boosting}
\title{Estimate component-wise boosting for dynamic models.}
\usage{
boosting(x, y, v = 0.2, minIt = ncol(x)/2, maxIt = 10 * ncol(x),
  ic.break = TRUE)
}
\arguments{
\item{x}{Matrix of independent variables. Each row is an observation and each column is a variable.}

\item{y}{Response variable equivalent to the function.}

\item{v}{Algorithm step size.}

\item{minIt}{Minimum number of iterations in case ic.break=TRUE.}

\item{maxIt}{Maximum number of iterations.}

\item{ic.break}{If TRUE, algorithm breaks when the minimum information criteria is likely to be found. If FALSE the algorithm stops only when maxIt is reached (default=TRUE).}
}
\value{
An object with S3 class boosting.
\item{coefficients}{Boosting coefficients for the model with the smallest information criteria.}
\item{fitted.values}{In-sample fitted values.}
\item{residuals}{Model residuals.}
\item{best.crit}{The smalles information criterion found.}
\item{crit}{The information criterion in each iteration.}
\item{df}{Degrees of freedom.}
\item{coef.selection.count}{How many times each variable was selected.}
\item{y}{The supplied y.}
\item{call}{The matched call.}
}
\description{
Estimates the Bai & Ng (2009) component-wise boosting for dynamic models.
}
\details{
This is an implementation of time-series component-wise boosting using the results from Bai and Ng (2009). The algorithm has its own way of determining when to stop. Keep ic.break=TRUE if you want to use the standard stop criterion based on information criterion.

Note that the information criterion automaticaly adjusts the degrees of freedom of the model considering that the boosting may select the same variable more than once.
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##
data("BRinf")

## == Data preparation == ##
## == The model is yt = a + Xt-1'b + ut == ##
aux = embed(BRinf,2)
y=aux[,1]
x=aux[,-c(1:ncol(BRinf))]

## == Use factors == ##
factors=prcomp(x,scale. = TRUE)
xfact=factors$x[,1:10]

model=boosting(xfact,y)
coef(model)

plot(y,type="l")
lines(fitted(model),col=2)

}
\references{
Bai, Jushan, and Serena Ng. "Boosting diffusion indices." Journal of Applied Econometrics 24.4 (2009): 607-629.

Garcia, Medeiros and Vasconcelos (2017).
}
\seealso{
\code{\link{predict.boosting}}
}
\keyword{Boosting,}
\keyword{Dynamic}
\keyword{models}
