% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BRinf}
\alias{BRinf}
\title{A dataset used in the examples}
\format{A matrix with 156 rows and 92 variables.}
\usage{
data(BRinf)
}
\description{
This data is from Garcia, Medeiros and Vasconcelos (2017). It has 93 variables which are divided between inflation indexes, macroeconomic variables and expectation variables.
}
\references{
Garcia, Medeiros and Vasconcelos (2017).
}
\keyword{datasets}
