% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDvar.R
\name{HDvar}
\alias{HDvar}
\title{Estimate High-dimensional VARs}
\usage{
HDvar(Y, p, fn = NULL, xreg = NULL)
}
\arguments{
\item{Y}{Time-series matrix or data.frame with the VAR endogenous variables.}

\item{p}{Lag order (default = 1).}

\item{fn}{A function that receives x (independent variables) and y (dependent variable) and returns the model coefficients (default is OLS).}

\item{xreg}{Exogenous controls.}
}
\value{
An object with S3 class "HDeconometricsVAR", "HDvar".
\item{coef.by.equation}{Coefficients listed by each VAR equation.}
\item{coef.by.block}{Coefficients separated by blocks (intercepts, lags, exogenous).}
\item{fitted.values}{In-sample fitted values.}
\item{residuals}{The residuals.}
\item{Y}{Supplied endogenous data.}
\item{p}{VAR lag order chosen by the user.}
\item{N}{Number of endogenous variables.}
\item{covmat}{Residuals covariance matrix.}
\item{xreg}{Exogenous controls supplied by the user.}
\item{T}{Number of observations.}
\item{fn}{The estimation function.}
\item{call}{The matched call.}
}
\description{
Estimate High-dimensional Vector Autorregressive models using the model chosen by the user. The model must be parametric.
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##
data("BRinf")
Y=BRinf[,1:59]# remove expectation variables

fn=function(x,y){
  model=ic.glmnet(x,y)
  coef(model)
}

modelH=HDvar(Y,p=4,fn=fn)

# take a look at the coefficients
eq=coef(modelH,type="equation")
block=coef(modelH,type="block")
block$Lag1


}
\references{
Garcia, Medeiros and Vasconcelos (2017).
}
\seealso{
\code{\link{predict}}, \code{\link{lbvar}}, \code{\link{irf}}
}
\keyword{Bayesian}
\keyword{High-dimension,}
\keyword{LASSO,}
\keyword{VAR,}
\keyword{models}
