% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{identification}
\alias{identification}
\title{Recursive identification of a VAR model}
\usage{
identification(object, type = c("recursive", "identity"))
}
\arguments{
\item{object}{A HDvar or lbvar object.}

\item{type}{"recursive" or "identity". The last is just using a identity matrix as contemporaneous shocks.}
}
\value{
A list with the following objects:
\item{A}{Matrix of contemporaneous coefficients.}
\item{sigma2u}{Structural shocks covariance matrix.}
\item{type}{Matched type.}
}
\description{
Estimates the contemporaneous coefficients and the structural shocks covariance matrix using recursive identification or using a identity matrix.
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##

# = This is an ilustrative example = #
# = The identification ignores which variables are more exogenous = #

data("BRinf")
Y=BRinf[,1:59]# remove expectation variables
modelB=lbvar(Y,p=4)
identB=identification(modelB)

}
\references{
Garcia, Medeiros and Vasconcelos (2017).
}
