% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{VAR ipulse response functions}
\usage{
irf(object, ident, h, boot = FALSE, M = 100, unity.shock = TRUE)
}
\arguments{
\item{object}{A HDvar or lbvar object.}

\item{ident}{A list with two elements: $A must be a matrix with the contemporaneus coefficients of a identified VAR and $sigma2u must be the structural shocks covariance matrix. This is the output of the function chol.identification.}

\item{h}{Number of steps ahead.}

\item{boot}{If TRUE bootstrap confidence bands are calculated (default=FALSE).}

\item{M}{Number of bootstrap replications}

\item{unity.shock}{If TRUE the impulses are equal 1. If FALSE the impulses are of one standard deviation (default=TRUE).}
}
\value{
An object with S3 class "irf".
\item{point.irf}{A list with the point ir coefficients. Each element in the list is a matrix with the response on all variables cause by an impulse on the variable that gives name to the matrix.}
\item{density}{Returned only if boot=TRUE. A list that stores the boostrap ir coefficients. Each element in the list is another list with the response on all variables cause by an impulse on the variable that gives name to the list.}
}
\description{
Estimates impulse response coefficients of a HDvar or lbvar model h steps ahead. Confidence bands may be computed by bootstrap.
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##

# = This is an ilustrative example = #
# = The identification ignores which variables are more exogenous = #
data("BRinf")
Y=BRinf[,1:59]# remove expectation variables
modelB=lbvar(Y,p=3)
identB=identification(modelB)
irfB=irf(modelB,identB,h=12,boot = TRUE,M=100)
plot(irfB,1,2,alpha=0.1)


}
\references{
Garcia, Medeiros and Vasconcelos (2017).
}
\seealso{
\code{\link{predict}}, \code{\link{lbvar}}, \code{\link{identification}}, \code{\link{plot.irf}}
}
\keyword{Bayesian}
\keyword{High-dimension,}
\keyword{VAR,}
\keyword{irf,}
\keyword{models}
