% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbvar.R
\name{lbvar}
\alias{lbvar}
\title{Estimate Large Bayesian VARs}
\usage{
lbvar(Y, p = 1, delta = 0, lambda = 0.05, xreg = NULL)
}
\arguments{
\item{Y}{Time-series matrix or data.frame with the VAR endogenous variables.}

\item{p}{Lag order (default = 1).}

\item{delta}{Numeric vector indicating the prior for the autorregressive coefficients (default = 0 for all variables). If the prior is the same for all variables the user may supply a single number. Otherwise the vector must have one element for each variable.}

\item{lambda}{Constant that regulates the importance given to the priors (default = 0.05). If lambda = 0 the model ignores the data and the posterior equal the prior. For bigger lambda the model converges to the OLS  estimates.}

\item{xreg}{Exogenous controls.}
}
\value{
An object with S3 class "HDeconometricsVAR", "lbvar".
\item{coef.by.equation}{Coefficients listed by each VAR equation.}
\item{coef.by.block}{Coefficients separated by blocks (intercepts, lags, exogenous).}
\item{fitted.values}{In-sample fitted values.}
\item{residuals}{The residuals.}
\item{Y}{Supplied endogenous data.}
\item{p}{VAR lag order chosen by the user.}
\item{N}{Number of endogenous variables.}
\item{covmat}{Residuals covariance matrix.}
\item{xreg}{Exogenous controls supplied by the user.}
\item{Ts}{Number of real observations, number of dummy observations and the sum of both.}
\item{delta}{The delta chosen.}
\item{lambda}{The lambda chosen.}
\item{call}{The matched call.}
}
\description{
Estimate large Bayesian Vector Autorregressive models from Banbura et al. (2010)
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##
data("BRinf")
Y=BRinf[,1:59]# remove expectation variables
modelB=lbvar(Y,p=4)

# take a look at the coefficients
eq=coef(modelB,type="equation")
block=coef(modelB,type="block")
block$Lag1

}
\references{
Banbura, M., Giannone, D., & Reichlin, L. (2010). Large Bayesian vector autoregressions. Journal of Applied Econometrics, 25, 71–92.

Garcia, Medeiros and Vasconcelos (2017).
}
\seealso{
\code{\link{predict}}, \code{\link{HDvar}}, \code{\link{irf}}, \code{\link{fitLambda}}
}
\keyword{BVAR,}
\keyword{Bayesian}
\keyword{High-dimension,}
\keyword{VAR,}
\keyword{models}
