% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ic.glmnet.R
\name{plot.ic.glmnet}
\alias{plot.ic.glmnet}
\title{Plot the information criterion curve through the regularization path}
\usage{
\method{plot}{ic.glmnet}(x, ...)
}
\arguments{
\item{x}{A ic.glmnet object estimated using the ic.glmnet function.}

\item{...}{Arguments to be passed to other methods.}
}
\description{
Plots the information criterion curve as a function of log(lambda) through the regularization path.
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##
data("BRinf")

## == Data preparation == ##
## == The model is yt = a + Xt-1'b + ut == ##
aux = embed(BRinf,2)
y=aux[,1]
x=aux[,-c(1:ncol(BRinf))]

## == LASSO == ##
lasso=ic.glmnet(x,y,crit = "bic")
plot(lasso)

}
