% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.boosting.R
\name{predict.boosting}
\alias{predict.boosting}
\title{Predict method for boosting objects}
\usage{
\method{predict}{boosting}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A boosting object estimated using the ic.glmnet function.}

\item{newdata}{An optional data to look for the explanatory variables used to predict. If omitted, the fitted values are used.}

\item{...}{Arguments to be passed to other methods.}
}
\description{
Predicted values based on boosting object.
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##
data("BRinf")

## == Data preparation == ##
## == The model is yt = a + Xt-1'b + ut == ##
aux = embed(BRinf,2)
y=aux[,1]
x=aux[,-c(1:ncol(BRinf))]

## == break data (in-sample and out-of-sample)
yin=y[1:120]
yout=y[-c(1:120)]
xin=x[1:120,]
xout=x[-c(1:120),]

## == Use factors == ##
factors=prcomp(xin,scale. = TRUE)
xfact=factors$x[,1:10]

model=boosting(xfact,yin)

xfactout=predict(factors,xout)[,1:10]
pred=predict(model,xfactout)

plot(yout,type="l")
lines(pred,col=2)
}
