% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ic.glmnet.R
\name{predict.ic.glmnet}
\alias{predict.ic.glmnet}
\title{Predict method for ic.glmnet objects}
\usage{
\method{predict}{ic.glmnet}(object, newdata, ...)
}
\arguments{
\item{object}{A ic.glmnet object estimated using the ic.glmnet function.}

\item{newdata}{An optional data to look for the explanatory variables used to predict. If omitted, the fitted values are used.}

\item{...}{Arguments to be passed to other methods.}
}
\description{
Predicted values based on ic.glmnet object.
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##
data("BRinf")

## == Data preparation == ##
## == The model is yt = a + Xt-1'b + ut == ##
aux = embed(BRinf,2)
y=aux[,1]
x=aux[,-c(1:ncol(BRinf))]

## == Break the data into in-sample and out-of-sample
y.in=y[1:100]; y.out=y[-c(1:100)]
x.in=x[1:100,]; x.out=x[-c(1:100),]

## == LASSO == ##
lasso=ic.glmnet(x.in,y.in,crit = "bic")

pred=predict(lasso,newdata=x.out)
plot(y.out, type="l")
lines(pred, col=2)

}
