% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{voldata}
\alias{voldata}
\title{A dataset used in the examples}
\format{A dataframe with 5521 rows and 30 variables.}
\usage{
data(voldata)
}
\description{
This is a panel of volatilities with daily data for the 30 Dow stocks estimated using a garch(1,1) for each serie. The returns used on the garch are from the dataset dji30ret from te package rugarch.
}
\references{
Alexios Ghalanos (2015). rugarch: Univariate GARCH models. R package version 1.3-6.
}
\keyword{datasets}
