% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{identification}
\alias{identification}
\title{Recursive identification of a VAR model}
\usage{
identification(object)
}
\arguments{
\item{object}{A HDvar or lbvar object.}
}
\value{
A list with the following objects:
\item{A}{Matrix of contemporaneous coefficients.}
\item{sigma2u}{Structural shocks covariance matrix.}
}
\description{
Estimates the contemporaneous coefficients and the structural shocks covariance matrix using recursive identification.
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##

# = This is an ilustrative example = #
# = The identification ignores which variables are more exogenous = #

data("BRinf")
Y=BRinf[,1:59]# remove expectation variables
modelB=lbvar(Y,p=4)
identB=identification(modelB)

}
\references{
Garcia, Medeiros and Vasconcelos (2017).
}
